# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Bike:
    properties:
      id:
        type: integer
      is_available:
        type: boolean
      latitude:
        type: number
      longitude:
        type: number
      price_per_minute:
        type: number
    type: object
  models.Rental:
    properties:
      bike_id:
        type: integer
      cost:
        type: number
      created_at:
        type: string
      duration_minutes:
        type: integer
      end_latitude:
        type: number
      end_longitude:
        type: number
      end_time:
        type: string
      id:
        type: integer
      start_latitude:
        type: number
      start_longitude:
        type: number
      start_time:
        type: string
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      updated_at:
        type: string
    type: object
...
  types.AddBikeRequest:
    properties:
      latitude:
        type: number
      longitude:
        type: number
      price_per_minute:
        type: number
    type: object
...
  types.ErrorResponse:
    properties:
      details:
        additionalProperties:
          type: string
        type: object
      error:
        type: string
    type: object
...
  types.PaginatedResponse:
    properties:
      data: {}
      limit:
        type: integer
      message:
        type: string
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  types.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
  types.UpdateBikeRequest:
    properties:
      is_available:
        type: boolean
      latitude:
        type: number
      longitude:
        type: number
      price_per_minute:
        type: number
    type: object
  types.UpdateRentalRequest:
    properties:
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@bikerental.com
    name: API Support
  description: API para gestión de alquiler de bicicletas
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Bike Rental Service API
  version: "1.0"
paths:
  /admin/bikes:
    get:
      consumes:
      - application/json
      description: Get paginated list of all bikes in the system (requires admin authentication)
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page (max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: All bikes retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/types.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Bike'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Get all bikes (Admin)
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: Create a new bike with location and pricing (requires admin authentication)
      parameters:
      - description: Bike data with coordinates and price
        in: body
        name: bike
        required: true
        schema:
          $ref: '#/definitions/types.AddBikeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Bike added successfully
          schema:
            allOf:
            - $ref: '#/definitions/types.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Bike'
              type: object
        "400":
          description: Invalid coordinates or price
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "401":
          description: Unauthorized - admin credentials required
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Add a new bike (Admin)
      tags:
      - admin
  /admin/bikes/{bike-id}:
    patch:
      consumes:
      - application/json
      description: Update bike details like location, availability, or price (requires
        admin authentication)
      parameters:
      - description: Bike ID
        in: path
        name: bike-id
        required: true
        type: integer
      - description: Bike update data
        in: body
        name: bike
        required: true
        schema:
          $ref: '#/definitions/types.UpdateBikeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Bike updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/types.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Bike'
              type: object
        "400":
          description: Invalid bike ID or update data
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "404":
          description: Bike not found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Update bike (Admin)
      tags:
      - admin
  /admin/rentals:
    get:
      consumes:
      - application/json
      description: Get paginated list of all rentals in the system (requires admin
        authentication)
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page (max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: All rentals retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/types.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Rental'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Get all rentals (Admin)
      tags:
      - admin
  /admin/rentals/{rental-id}:
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific rental (requires admin
        authentication)
      parameters:
      - description: Rental ID
        in: path
        name: rental-id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Rental details retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/types.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Rental'
              type: object
        "400":
          description: Invalid rental ID
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "404":
          description: Rental not found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Get rental details (Admin)
      tags:
      - admin
    patch:
      consumes:
      - application/json
      description: Update rental status (requires admin authentication)
      parameters:
      - description: Rental ID
        in: path
        name: rental-id
        required: true
        type: integer
      - description: Rental update data with status
        in: body
        name: rental
        required: true
        schema:
          $ref: '#/definitions/types.UpdateRentalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Rental updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/types.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Rental'
              type: object
        "400":
          description: Invalid rental ID or status required
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Update rental (Admin)
      tags:
      - admin
  /admin/users:
    get:
      consumes:
      - application/json
      description: Get paginated list of all users in the system (requires admin authentication)
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page (max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: All users retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/types.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.User'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Get all users (Admin)
      tags:
      - admin
...
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"